add tiles to constants
